==> Lessons/06_Part2_InteractiveMarkdown.Rmd <==
---
title: '6 - Part 2: Interactive R Markdown'
author: Environmental Data Analytics | John Fay & Luana Lima 
date: "Fall 2023"
output: html_document
runtime: shiny
---

#ERROR: can not find '%>%' when exporting the document as html

#1.  add YAML

#2.  edit chunk

 #   1.  echo=FALSE to hide the running process, only showing the results

  #  2.  warning=FALSE

3.  list the context , tables, and figures

4.  add widges

\listoffigures

\listoftables

## Overview

RMarkdown is a powerful tool for creating dynamic documents that combine code, text, and graphics. With the addition of widgets, it's possible to make these documents even more interactive and engaging. In this lesson, we will explore how to create interactive HTML files using RMarkdown widgets.

## LESSON OBJECTIVES

1.  Create plotting functions that accept variables
2.  Use widgets from the `shiny` library to make interactive HTML documents

### Set up the coding environment

```{r install packages,warning=FALSE, echo=FALSE, include = FALSE }
#error message might occur when running the document 

#options(repos = "https://cran.rstudio.com/")
#install.packages("tidyverse")
#install.packages("lubridate")
#install.packages("here")
#install.packages("dplyr")
#install.packages("shiny")
#install.packages("ggplot2")


```

```{r setup, echo=FALSE, warning=FALSE, message=FALSE,eval=FALSE}
# error message: eval=FALSE, does not run the chunk 

#Load packages
library(tidyverse);library(here);library(lubridate);library(dplyr);library(shiny);library(ggplot2)

#Load data
nutrient_data <- read.csv(
  here('Data/Processed_KEY','NTL-LTER_Lake_Nutrients_PeterPaul_Processed.csv'),
  stringsAsFactors = TRUE)

#Fix dates
nutrient_data$sampledate = ymd(nutrient_data$sampledate)

#Set a default plotting theme
my_theme <- theme_classic(base_size = 14)
theme_set(my_theme) # do not have to recall "my_theme" in every plot chunk

```

### Create a base plot

We will plot Total P as a function of time, coloring data by depth, and symbolizing by lake

```{r Create base plot, echo=FALSE, eval=FALSE,fig.cap="Concentration over time, by depth and lake"}
#Construct the plot
nutrient_data %>% 
  filter(year4 > 1990 & year4 < 2005) %>% 
  ggplot(
    aes(
      x = sampledate, 
      y = tp_ug, 
      fill = depth_id, 
      shape = lakename)) +
  geom_point(alpha = 0.8, size = 2) +
  scale_x_date(
    date_breaks = "1 years",
    date_labels = '%Y') +
  scale_shape_manual(values = c(21, 24)) +
  labs(
    title = "Concentration over time, by depth and lake",
    x = "Date", 
    y = expression(Concentration ~ (mu*g / L)), 
    #use expression to print Greek letter mu
    shape = "Lake", 
    fill = "Depth ID") +
  scale_fill_viridis_c(
    option = "mako", 
    begin = 0, 
    end = 1, 
    guide = guide_colorbar(reverse = T), #reverse filled colors
    direction = -1)
```

## Using variables in plots

Now tweak the above code so that the variable we plot (`tp_ug`) is specified as a variable rather than hard coded in our ggplot command. We will also set the start and end dates to variables as well.

To do this we need to convince R that the value we associate with the `y` aesthetic is a variable, not a column name. This is done by the odd convention of inserting the variable within the code `!!sym()`. The `!!` (pronounced "bang-bang") combined with `sym()` are "Tidy eval helpers" necessary because tidy-R does away with quotations. See more info [here](https://ggplot2.tidyverse.org/reference/tidyeval.html).

```{r Tweak plot to use variables 1, echo=FALSE,fig.cap="Concentration over time, by depth and lake"}
the_parameter="tp_ug"
start_year=1990
end_year=2005

#Construct the plot
my_plot<-nutrient_data %>% 
  filter(year4 > start_year & year4 < end_year) %>% 
  ggplot(
    aes(
      x = sampledate, 
      y = !!sym(the_parameter), #to call out the parameter
      fill = depth_id, 
      shape = lakename)) +
  geom_point(alpha = 0.8, size = 2) +
  scale_x_date(
    date_breaks = "1 years",
    date_labels = '%Y') +
  scale_shape_manual(values = c(21, 24)) +
  labs(
    title = "Concentration over time, by depth and lake",
    x = "Date", 
    y = expression(Concentration ~ (mu*g / L)), 
    #use expression to print Greek letter mu
    shape = "Lake", 
    fill = "Depth ID") +
  scale_fill_viridis_c(
    option = "mako", 
    begin = 0, 
    end = 1, 
    guide = guide_colorbar(reverse = T), #reverse filled colors
    direction = -1)
my_plot
```

```{r Tweak plot to use variables 2, echo=FALSE,fig.cap="Concentration over time, by depth and lake"}

the_parameter="tp_ug"
start_year=1990
end_year=2005

#Create plotting function
plot_it<- function(the_parameter,start_year,end_year){
  #construct the plot 
  test_plot<-nutrient_data %>% 
    filter(year4 > start_year & year4 < end_year) %>% 
    ggplot(
      aes(
        x = sampledate, 
        y = !!sym(the_parameter), # to call out the parameter values 
        fill = depth_id, 
        shape = lakename)) +
    geom_point(alpha = 0.8, size = 2) +
    scale_x_date(
      date_breaks = "1 years",
      date_labels = '%Y') +
    scale_shape_manual(values = c(21, 24)) +
    labs(
      title = "Concentration over time, by depth and lake",
      x = "Date", 
      y = expression(Concentration ~ (mu*g / L)), 
      shape = "Lake", 
      fill = "Depth ID") +
    scale_fill_viridis_c(
      option = "mako", 
      begin = 0, 
      end = 1, 
      guide = guide_colorbar(reverse = T),
      direction = -1)
  return(test_plot)
}

plot_it("po4",1992,2005)
#plot_it: the name of the defioned function 
#plot_it(the parameter,start date, end date)

```

##Interactive R with Shiny Widgets The R MarkDown Cheat Sheet provides a short example on how to make your Markdown document interactive. First you must alter your yaml section to read:

```         
output: html_document
runtime: shiny
```

Notice that when you do so, the Knit menu is replaced.

Then we can add the code chunk listed in the document.Before running this code, however, we should set the code chunk options for the previous code chunks so that they are not run (eval) and not shown (echo).

```{r, echo=FALSE}
# numericInput: to create a widge that allows users to enter input 
numericInput(
  inputId="n",
  label="How many cars?",
  value=5)

# print out the results 
# input$the_name_for_inputId
renderPrint(paste("You chose",input$n))

# print thee first 5 rows of the cars data frame 
renderTable({head(cars,input$n)})

```

Shiny has many other input widgets: <https://shiny.rstudio.com/gallery/widget-gallery.html>. Below we'll use a Select Box widget to specifiy the variable shown in our plot. We'll also add a slider range to limit the date shown in our plot.

Note: Before running this plot, set the previous code chunk to not evaluate

```{r widget exploration, echo=FALSE}
# select box
selectInput(
  inputId="the parameter",#the name of the parameter
  label="Select the parameter to plot",# label will be shown on the screen
  choices=c("tp_ug","tn_ug","po4"), 
  selected="tp_ug") #the default option

# to generate a plot based on the input 
renderPlot(plot_it(input$the_parameter,1990,1993))

# slide range 
sliderInput(
  inputId="YearSlider",
  label="Select the date range",
  min=1990,#minimum value in the data frame 
  max=2020,
  value=c(1995,2000)) #default value

# print the results

renderPrint(input$YearSlider)

 # print the results
renderPlot(
  plot_it(input$the_parameter,
          input$input$YearSlider[1], #return the first value in YearSlider
          input$input$YearSlider[2]))

```
