==> Lessons/06_Test_Exercise_Markdown.Rmd <==
---
title: "Example Document"
subtitle: "subtitle here"
author: "Ina Liao"
output: pdf_document
date: "2023-10-07"
toc: true #contents
---

YAML enclosed by three dashes - - -

\listoffigures


\newpage #to put a page break

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## R Markdown

### Subheader

#### Smaller sub-header

-   First bullet

-   Second bullet

    -   Subbullet

> This is a block quote

1.  step1
2.  step2

This is an R Markdown document. Markdown is a simple formatting syntax for authoring HTML, PDF, and MS Word documents. For more details on using R Markdown see [[my link](http://rmarkdown.rstudio.com)] .

[impage link](img/rmarkdownflow.png) #might only work in HTML output

When you click the **Knit** button a document will be generated that includes both content as well as the output of any embedded R code chunks within the document. You can embed an R code chunk like this:

```{r cars}
summary(cars)
```

## Including Plots

You can also embed plots, for example:

```{r pressure, echo=FALSE,fig.cap="Temperature Distribution"}
plot(pressure)
```

Note that the `echo = FALSE` parameter was added to the code chunk to prevent printing of the R code that generated the plot.
